Rename "MarchingCubesPainter" to "MarchingCubes"
