+++ source/marchingcubes/CMakeLists.txt
@@ -41,7 +41,7 @@ set(sources
     ${source_path}/ChunkFactory.cpp
     ${source_path}/LookUpData.cpp
     ${source_path}/ManageChunksStage.cpp
-    ${source_path}/MarchingCubesPainter.cpp
+    ${source_path}/MarchingCubes.cpp
     ${source_path}/MarchingCubesPipeline.cpp
     ${source_path}/RenderStage.cpp
     ${source_path}/TerrainModificationStage.cpp
@@ -54,7 +54,7 @@ set(api_includes
     ${include_path}/ChunkFactory.h
     ${include_path}/LookUpData.h
     ${include_path}/ManageChunksStage.h
-    ${include_path}/MarchingCubesPainter.h
+    ${include_path}/MarchingCubes.h
     ${include_path}/MarchingCubesPipeline.h
     ${include_path}/RenderStage.h
     ${include_path}/TerrainModificationStage.h
